Checkpoint gds/shmem work.
Back to operational on my laptop.
