+++ src/mca/gds/shmem/gds_shmem_component.c
@@ -72,7 +72,8 @@ pmix_gds_shmem_component_t pmix_mca_gds_shmem_component = {
         .pmix_mca_query_component = component_query,
         .reserved = {0}
     },
-    .jobs = PMIX_LIST_STATIC_INIT
+    .jobs = PMIX_LIST_STATIC_INIT,
+    .sessions = PMIX_LIST_STATIC_INIT
 };
 
 /*
